refactored Home delegates
